refactor: Component hierarchy and code cleanup
