--- ui/desktop/helper-scripts/README.md
@@ -1,5 +1,0 @@
-Put `goosey` in your $PATH if you want to launch via:
-
-`goosey .`
-
-Will open goose GUI from any path you specify
