+++ ui/desktop/scripts/README.md
@@ -1,0 +1,9 @@
+# Goosey
+
+Put `goosey` in your $PATH if you want to launch via:
+
+```
+goosey .
+```
+
+This will open goose GUI from any path you specify
