+++ ui/desktop/scripts/add-macos-cert.sh
