+++ ui/desktop/src/components/MoreMenu.tsx
@@ -1,14 +1,14 @@
 import { Popover, PopoverContent, PopoverTrigger, PopoverPortal } from '@radix-ui/react-popover';
 import React, { useEffect, useState } from 'react';
 import { More } from './icons';
-import type { View } from '../../ChatWindow';
+import type { View } from '../ChatWindow';
 
 interface VersionInfo {
   current_version: string;
   available_versions: string[];
 }
 
-// Accept setView as a prop from the parent (e.g. ChatContent)
+// Accept setView as a prop from the parent (e.g. Chat)
 export default function MoreMenu({ setView }: { setView: (view: View) => void }) {
   const [open, setOpen] = useState(false);
   const [versions, setVersions] = useState<VersionInfo | null>(null);
