--- ui/desktop/test.json
@@ -1,13 +1,0 @@
-{
-    "messages": [
-      {
-        "role": "user",
-        "content": [
-          {
-            "type": "text",
-            "text": "Hello! ping www.apple.com 10 times."
-          }
-        ]
-      }
-    ]
-  }
